\name{gvcm.cat-internal}
\alias{a.coefs}
\alias{abc.a.coefs}
\alias{abc.optimierung}
\alias{check.simulation}
\alias{contr.effect}
\alias{cv.lambda.phi}
\alias{cv.vectors}
\alias{first.derivative}
\alias{lambda.max}
\alias{optimierung}
\alias{p}
\alias{path.matrix}
\alias{reduce}
\alias{v}
\alias{weight.function}
\title{Internal Function of gvcm.cat()}
\description{
For internal use only.
}
\seealso{
Function \code{\link{gvcm.cat}}
}
\keyword{gvcm.cat}
