\name{gvcm.cat-internal}
\alias{abcfit}
\alias{a.coefs}
\alias{abc.a.coefs}
\alias{bootstrap}
\alias{check.simulation}
\alias{contr.effect}
\alias{elastic}
\alias{cv.lambda}
\alias{cv.vectors}
\alias{grouped}
\alias{grouped.fused}
\alias{gvcmcatfit}
\alias{p}
\alias{pspline}
\alias{path.matrix}
\alias{reduce}
\alias{SCAD}
\alias{sp}
\alias{v}
\alias{vspline}
\alias{weight.function}
\title{Internal Function of gvcm.cat()}
\description{
For internal use only.
}
\seealso{
Function \code{\link{gvcm.cat}}
}
